Update assignment.js
Challenge -1: With the use of for-loop I was able to iterate the given array of numbers and sum it up while logging out the sum and the length of the array.

Challenge - 2 :  To count the even numbers I made use of the for-loop too and an if-statement. The modulus basically just prints the remainder of the division and if the remainder is equal to 0, it is simply an even number then the count++ increases the number